Improved Lua-based persistence API. Written files are now more clean * Added default values to GenericProperty<T> * Lua export now checks properties for default values and only sets the property if its value is different from the default value.

+++ core/pipeline/abstractpipeline.h
@@ -171,7 +171,7 @@ namespace campvis {
          * \param   index    The index of the processor to get
          * \return  The first processor whose name matches \a name, 0 if no such processor exists.
          */
-        AbstractProcessor* getProcessor(int index) const;
+        AbstractProcessor* getProcessor(size_t index) const;
 
         /**
          * Gets the flag whether this pipeline is currently enabled.
